Extract scroll behavior from the full flow test
